bird2: replace gcc flag workaround with patch fixing alignment issues
Previously, we used the -mno-unaligned-access flag to instruct GCC to
forbid unaligned memory access on arm processors. This was a workaround
for alignment issues that caused crashes.

This commit imports a patch from the BIRD mailing list discussion [0]
that resolves the alignment issue by modifying the net_addr structure,
removing the need for the GCC flag. The patch addresses the alignment
problem more efficiently and avoids the performance, code size, and
hardware optimization drawbacks of the flag.

[0] - http://trubka.network.cz/pipermail/bird-users/2024-December/017957.html

Signed-off-by: Nick Hainke <vincent@systemli.org>